door-key-open-detection-with-tflite/app/requirements.txt
pillow==9.3.0
tflite-runtime
numpy==1.19.5
